pyproject.toml
[tool.poetry]
name = "whisper-auto-subtitle-generator"
version = "0.1.0"
description = ""
authors = ["Serotina <wlsdn2749@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytube = "^15.0.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
ffmpeg-python = "^0.2.0"
anthropic = "^0.19.1"
tqdm = "^4.66.2"
transformers = "^4.38.2"
accelerate = "^0.27.2"
huggingface-hub = "^0.21.4"
deepl = "^1.17.0"
googletrans-py = {git = "https://github.com/StarkBotsIndustries/googletrans.git"}
uvicorn = {extras = ["standard"], version = "^0.28.0"}
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
google-generativeai = "^0.4.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
